+++ sdk/action_status/client.go
@@ -3,7 +3,7 @@
 // ActionStatus client
 //
 // Command:
-// $ goa gen github.com/rightscale/governance/front_service/design
+// $ goa gen github.com/rightscale/governance/front_service/design -o .
 
 package actionstatus