+++ sdk/applied_policy/client.go
@@ -3,7 +3,7 @@
 // AppliedPolicy client
 //
 // Command:
-// $ goa gen github.com/rightscale/governance/front_service/design
+// $ goa gen github.com/rightscale/governance/front_service/design -o .
 
 package appliedpolicy