+++ sdk/archived_incident/client.go
@@ -3,7 +3,7 @@
 // ArchivedIncident client
 //
 // Command:
-// $ goa gen github.com/rightscale/governance/front_service/design
+// $ goa gen github.com/rightscale/governance/front_service/design -o .
 
 package archivedincident