+++ sdk/archived_incident/views/view.go
@@ -3,7 +3,7 @@
 // ArchivedIncident views
 //
 // Command:
-// $ goa gen github.com/rightscale/governance/front_service/design
+// $ goa gen github.com/rightscale/governance/front_service/design -o .
 
 package views