+++ sdk/http/incident/client/client.go
@@ -3,7 +3,7 @@
 // Incident client HTTP transport
 //
 // Command:
-// $ goa gen github.com/rightscale/governance/front_service/design
+// $ goa gen github.com/rightscale/governance/front_service/design -o .
 
 package client