+++ sdk/http/policy_template/client/cli.go
@@ -3,7 +3,7 @@
 // PolicyTemplate HTTP client CLI support package
 //
 // Command:
-// $ goa gen github.com/rightscale/governance/front_service/design
+// $ goa gen github.com/rightscale/governance/front_service/design -o .
 
 package client