+++ sdk/http/published_template/client/encode_decode.go
@@ -3,7 +3,7 @@
 // PublishedTemplate HTTP client encoders and decoders
 //
 // Command:
-// $ goa gen github.com/rightscale/governance/front_service/design
+// $ goa gen github.com/rightscale/governance/front_service/design -o .
 
 package client