+++ sdk/http/published_template/client/paths.go
@@ -3,7 +3,7 @@
 // HTTP request path constructors for the PublishedTemplate service.
 //
 // Command:
-// $ goa gen github.com/rightscale/governance/front_service/design
+// $ goa gen github.com/rightscale/governance/front_service/design -o .
 
 package client