+++ sdk/incident_aggregate/endpoints.go
@@ -3,7 +3,7 @@
 // IncidentAggregate endpoints
 //
 // Command:
-// $ goa gen github.com/rightscale/governance/front_service/design
+// $ goa gen github.com/rightscale/governance/front_service/design -o .
 
 package incidentaggregate