+++ sdk/incident_aggregate/service.go
@@ -3,7 +3,7 @@
 // IncidentAggregate service
 //
 // Command:
-// $ goa gen github.com/rightscale/governance/front_service/design
+// $ goa gen github.com/rightscale/governance/front_service/design -o .
 
 package incidentaggregate